Working on Applying App Factory Pattern
