+++ application/endpoints/Auxiliary.py
