+++ application/endpoints/RateStudy.py
