+++ application/endpoints/RemoveWishlist.py
